+++ examples/send_e2e_text.rs
@@ -41,7 +41,7 @@ fn main() {
         process::exit(1);
     });
     let encrypted = api.encrypt_text_msg(&text, &recipient_key);
-    let msg_id = api.send(&to, &encrypted);
+    let msg_id = api.send(&to, &encrypted, false);
 
     match msg_id {
         Ok(id) => println!("Sent. Message id is {}.", id),